Extract EscapeSequenceImpl
